scd-proiect/api-gateway/requirements.txt
flask
sqlalchemy
psycopg2-binary
redis
requests
python-keycloak
